feat: adds login button to landing page (#68)
